Working on Videoalpha test fix.
The tests were failing because XMLDescriptor adds in some attributes
to _model_data, such as `xml_attributes`, that aren't necessary. The
solution is to handle all XML parsing in VideoDescriptor. There's
still one test failing, which has to do with the `url_name` field of a
pointer tag. I'm not sure what the expected behavior of url_name is,
but this fixes the main issues.